api/docker,volplugin: new api/docker package for storing docker api top-level handlers.
