+++ volplugin/handlers.go
@@ -11,80 +11,13 @@ import (
 	log "github.com/Sirupsen/logrus"
 	"github.com/contiv/errored"
 	"github.com/contiv/volplugin/api"
+	"github.com/contiv/volplugin/api/docker"
 	"github.com/contiv/volplugin/config"
 	"github.com/contiv/volplugin/errors"
 	"github.com/contiv/volplugin/lock"
 	"github.com/contiv/volplugin/storage"
 	"github.com/contiv/volplugin/storage/backend"
-	"github.com/docker/docker/pkg/plugins"
 )
-
-type unmarshalledConfig struct {
-	Request api.VolumeCreateRequest
-	Name    string
-	Policy  string
-}
-
-func (dc *DaemonConfig) nilAction(w http.ResponseWriter, r *http.Request) {
-	content, err := json.Marshal(api.VolumeCreateResponse{})
-	if err != nil {
-		api.DockerHTTPError(w, errors.UnmarshalRequest.Combine(err))
-		return
-	}
-	w.Write(content)
-}
-
-func (dc *DaemonConfig) activate(w http.ResponseWriter, r *http.Request) {
-	content, err := json.Marshal(plugins.Manifest{Implements: []string{"VolumeDriver"}})
-	if err != nil {
-		api.DockerHTTPError(w, errors.MarshalResponse.Combine(err))
-		return
-	}
-
-	w.Write(content)
-}
-
-func (dc *DaemonConfig) deactivate(w http.ResponseWriter, r *http.Request) {
-	w.WriteHeader(200)
-}
-
-func unmarshalAndCheck(w http.ResponseWriter, r *http.Request) (*unmarshalledConfig, error) {
-	defer r.Body.Close()
-	vr, err := unmarshalRequest(r.Body)
-	if err != nil {
-		api.DockerHTTPError(w, errors.UnmarshalRequest.Combine(err))
-		return nil, err
-	}
-
-	if vr.Name == "" {
-		api.DockerHTTPError(w, errors.InvalidVolume.Combine(errored.New("Name is empty")))
-		return nil, err
-	}
-
-	policy, name, err := splitPath(vr.Name)
-	if err != nil {
-		api.DockerHTTPError(w, errors.ConfiguringVolume.Combine(err))
-		return nil, err
-	}
-
-	uc := unmarshalledConfig{
-		Request: vr,
-		Name:    name,
-		Policy:  policy,
-	}
-
-	return &uc, nil
-}
-
-func writeResponse(w http.ResponseWriter, vr *api.VolumeCreateResponse) {
-	content, err := json.Marshal(*vr)
-	if err != nil {
-		api.DockerHTTPError(w, errors.MarshalResponse.Combine(err))
-		return
-	}
-
-	w.Write(content)
-}
 
 func (dc *DaemonConfig) get(w http.ResponseWriter, r *http.Request) {
 	defer r.Body.Close()
@@ -134,7 +67,7 @@ func (dc *DaemonConfig) get(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	do, err := dc.volumeToDriverOptions(volConfig)
+	do, err := volConfig.ToDriverOptions(dc.Global.Timeout)
 	if err != nil {
 		api.DockerHTTPError(w, errors.MarshalVolume.Combine(err))
 		return
@@ -168,7 +101,7 @@ func (dc *DaemonConfig) list(w http.ResponseWriter, r *http.Request) {
 	for _, volume := range volList {
 		parts := strings.SplitN(volume, "/", 2)
 		if len(parts) != 2 {
-			log.Errorf("")
+			log.Errorf("Invalid volume %q detected iterating volumes", volume)
 			continue
 		}
 		if volObj, err := dc.Client.GetVolume(parts[0], parts[1]); err != nil {
@@ -184,7 +117,7 @@ func (dc *DaemonConfig) list(w http.ResponseWriter, r *http.Request) {
 			return
 		}
 
-		do, err := dc.volumeToDriverOptions(volConfig)
+		do, err := volConfig.ToDriverOptions(dc.Global.Timeout)
 		if err != nil {
 			api.DockerHTTPError(w, errors.MarshalVolume.Combine(err))
 			return
@@ -209,7 +142,7 @@ func (dc *DaemonConfig) list(w http.ResponseWriter, r *http.Request) {
 }
 
 func (dc *DaemonConfig) getPath(w http.ResponseWriter, r *http.Request) {
-	uc, err := unmarshalAndCheck(w, r)
+	uc, err := docker.Unmarshal(r)
 	if err != nil {
 		api.DockerHTTPError(w, errors.UnmarshalRequest.Combine(err))
 		return
@@ -231,7 +164,7 @@ func (dc *DaemonConfig) getPath(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	writeResponse(w, &api.VolumeCreateResponse{Mountpoint: path})
+	docker.WriteResponse(w, &api.VolumeCreateResponse{Mountpoint: path})
 }
 
 func (dc *DaemonConfig) returnMountPath(w http.ResponseWriter, driver storage.MountDriver, driverOpts storage.DriverOptions) {
@@ -241,11 +174,11 @@ func (dc *DaemonConfig) returnMountPath(w http.ResponseWriter, driver storage.Mo
 		return
 	}
 
-	writeResponse(w, &api.VolumeCreateResponse{Mountpoint: path})
+	docker.WriteResponse(w, &api.VolumeCreateResponse{Mountpoint: path})
 }
 
 func (dc *DaemonConfig) mount(w http.ResponseWriter, r *http.Request) {
-	uc, err := unmarshalAndCheck(w, r)
+	uc, err := docker.Unmarshal(r)
 	if err != nil {
 		api.DockerHTTPError(w, errors.UnmarshalRequest.Combine(err))
 		return
@@ -268,11 +201,11 @@ func (dc *DaemonConfig) mount(w http.ResponseWriter, r *http.Request) {
 			log.Warnf("Duplicate mount of %q detected: Lock failed", volName)
 			api.DockerHTTPError(w, errors.LockFailed.Combine(err))
 			return
-		} else {
-			log.Warnf("Duplicate mount of %q detected: returning existing mount path", volName)
-			dc.returnMountPath(w, driver, driverOpts)
-			return
 		}
+
+		log.Warnf("Duplicate mount of %q detected: returning existing mount path", volName)
+		dc.returnMountPath(w, driver, driverOpts)
+		return
 	}
 
 	ut := &config.UseMount{
@@ -330,11 +263,11 @@ func (dc *DaemonConfig) mount(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	writeResponse(w, &api.VolumeCreateResponse{Mountpoint: path})
+	docker.WriteResponse(w, &api.VolumeCreateResponse{Mountpoint: path})
 }
 
 func (dc *DaemonConfig) unmount(w http.ResponseWriter, r *http.Request) {
-	uc, err := unmarshalAndCheck(w, r)
+	uc, err := docker.Unmarshal(r)
 	if err != nil {
 		api.DockerHTTPError(w, errors.UnmarshalRequest.Combine(err))
 		return
@@ -382,28 +315,4 @@ func (dc *DaemonConfig) unmount(w http.ResponseWriter, r *http.Request) {
 	}
 
 	dc.returnMountPath(w, driver, driverOpts)
-}
-
-func (dc *DaemonConfig) capabilities(w http.ResponseWriter, r *http.Request) {
-	content, err := json.Marshal(map[string]map[string]string{
-		"Capabilities": {
-			"Scope": "global",
-		},
-	})
-
-	if err != nil {
-		api.DockerHTTPError(w, errors.UnmarshalRequest.Combine(err))
-		return
-	}
-
-	w.Write(content)
-}
-
-// Catchall for additional driver functions.
-func (dc *DaemonConfig) action(w http.ResponseWriter, r *http.Request) {
-	defer r.Body.Close()
-	log.Debugf("Unknown driver action at %q", r.URL.Path)
-	content, _ := ioutil.ReadAll(r.Body)
-	log.Debug("Body content:", string(content))
-	w.WriteHeader(503)
 }
